Implement basic plugin architecture
Can now use PV calculation functions stored in .py files in designated
plugin directories.  Moved all the existing pv calc functions over to
one of these files.

Also, added code to add the contents of the MANTIDPATH env varible to
the list of directories to search for modules.  If we can't import any
of the Mantid framework, print a nice error message and exit.