More packaging work
Changed the setup.py around to automatically generate its setup.cfg
file.  This is because we need setup.cfg to specify the 'prefix'
and install_lib options, but the setup.py script itself also needs to
know these values.  It turns out the init.d script also needs to know
these locations, so setup.py also reads a template for the init.d
script and modifies it.  All this means that install directories only
need to be set once in setup.py and the rpm will be built properly.

Note: The init script (and also the startup script in the bin directory)
still need some work.  The program doesn't actually start properly from
the init script yet.