More packaging changes
Changed setup.py so that the RPM installs the init file directly into
/etc/init.d and codes it to use
"/home/controls/<beamline>/mantidstats/mantidstats.conf" for the config
file.  The beamline name is determined from the hostname.

Also added postinstall and preuninstall scriptlets to the RPM to run
chkconfig and set the service to start on bootup.