Change how the live listener algorithm is started
Start the live listener algorithm with the option to reuse the existing
workspace (instead of creating a new one and saving the old one at
every run transition).  Also add some code to the start_live_listener
function to make sure that the MonitorLiveData algorithm is actually up
and running before returning from the function.  (It looks like we had a
case where we returned from start_live_listener before MonitorLiveData
was running, the watchdog loop caught that and tried to execute
start_live_data again.  That didn't work so well...)