Don't enable the PostProcessing algorithm if it's not needed
If the config file doesn't specify any PV's that require the
PostProcessing algorithm, then don't enable it in the call to
StartLiveData.  This also allows us to set PreserveEvents to False.
(Setting it to True was only necessary to force the workspace passed to
the PostProcessing alg to be an EventWorkspace.)  This will save quite a
bit of memory and also some CPU time.