Rip out Channel Access Server code
Replace the 'pcaspy' library with the 'pyepics' library.  Pyepics
doesn't do the server part, though, so that has to be handled by an
external softIOC program.  (Which is fine, because it looks very much
like pcaspy didn't do the server thing very well.)

This is the first step.  It's writing to the process variables (which is
the main thing), but the softIOC needs to be started mannually.  There's
some work to be done to make that step easier to automate.