Update startup scripts to use the softIoc binary
Added a script in bin to start the python with the proper command line
to auto generate the softIoc config files.  Also modified the existing
startup script to modify PYTHONPATH so that it can find the local epics
code.  Modified the initd template to generate the softIoc commands,
start softIoc and then start the main python code.  Also added code to
shut everything down properly.  Last, modified setyup.py to include all
the new files in the RPM.