Add event histogram plugin
First cut at a plugin to output event histograms.  Including a quick
python script to read the PV data and display an image.  Compared to what I
see in the MantidPlot instrument view, the data looks correct.  Need a
more precise test, though.

Note:  Since this is a waveform PV, the .db file generated by the
--generate_softioc_files command line option will be wrong!  It needs to
be modified by hand.  (Yes, this is something I intend to fix very
soon!)